requirements.txt
gradio==4.44.0
llama-index==0.10.0
llama-index-llms-fireworks==0.1.0
llama-index-embeddings-huggingface==0.2.0
pypdf==4.3.1
python-dotenv==1.0.1